Set up infra so we can move the wpt-manifest out of tree
Changes the wpt manifest path to the topobjdir instead so it can be moved out of tree.
Other changes so that the manifest download and update, and |mach wpt| and |mach test <wpt-test>| work with the new path.
The manifest is also downloaded and updated when creating the tests-archive to ensure that it exists when we run tests on TC.

Differential Revision: https://phabricator.services.mozilla.com/D5312

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1473915
gecko-commit: dd80de19f1429e0e00ec92e6491e25a2490ad99c
gecko-integration-branch: autoland
gecko-reviewers: gps